backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
py-algorand-sdk
pydantic
python-multipart
requests
motor
pymongo
